moved files around, created flow, wrote readme
