+++ .envrc
@@ -2,6 +2,7 @@
 [ -f .env ] && source .env
 
 # export S3_INPUT_PATH="s3://nesta-ds-projects/<MISSION OR PRACTICE>/<PROJECT NAME>"
+export METAFLOW_PROFILE=ds-cookiecutter
 
 # Activate the conda environment
 eval "$(conda shell.bash activate "dap_aria_mapping")"
